create init command
