+++ cmd/rollkit/commands/init.go
@@ -1,0 +1,94 @@
+package commands
+
+import (
+	"fmt"
+	"math/rand"
+	"time"
+
+	"cosmossdk.io/log"
+	cometp2p "github.com/cometbft/cometbft/p2p"
+	cometprivval "github.com/cometbft/cometbft/privval"
+	comettypes "github.com/cometbft/cometbft/types"
+	"github.com/spf13/cobra"
+)
+
+func NewInitNodeCmd() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "init",
+		Short: "Initialize the rollkit node",
+		RunE: func(cmd *cobra.Command, args []string) error {
+			err := parseConfig(cmd)
+			if err != nil {
+				return err
+			}
+
+			// use aggregator by default if the flag is not specified explicitly
+			if !cmd.Flags().Lookup("rollkit.aggregator").Changed {
+				nodeConfig.Aggregator = true
+			}
+
+			return initFiles(log.NewNopLogger())
+		},
+	}
+	return cmd
+}
+
+// TODO (Ferret-san): modify so that it initiates files with rollkit configurations by default
+// note that such a change would also require changing the cosmos-sdk
+func initFiles(logger log.Logger) error {
+	// Generate the private validator config files
+	cometprivvalKeyFile := config.PrivValidatorKeyFile()
+	cometprivvalStateFile := config.PrivValidatorStateFile()
+
+	var pv *cometprivval.FilePV
+	if fileExists(cometprivvalKeyFile) {
+		pv = cometprivval.LoadFilePV(cometprivvalKeyFile, cometprivvalStateFile)
+		logger.Info("Found private validator", "keyFile", cometprivvalKeyFile,
+			"stateFile", cometprivvalStateFile)
+	} else {
+		pv = cometprivval.GenFilePV(cometprivvalKeyFile, cometprivvalStateFile)
+		pv.Save()
+		logger.Info("Generated private validator", "keyFile", cometprivvalKeyFile,
+			"stateFile", cometprivvalStateFile)
+	}
+
+	// Generate the node key config files
+	nodeKeyFile := config.NodeKeyFile()
+	if fileExists(nodeKeyFile) {
+		logger.Info("Found node key", "path", nodeKeyFile)
+	} else {
+		if _, err := cometp2p.LoadOrGenNodeKey(nodeKeyFile); err != nil {
+			return err
+		}
+		logger.Info("Generated node key", "path", nodeKeyFile)
+	}
+
+	// Generate the genesis file
+	genFile := config.GenesisFile()
+	if fileExists(genFile) {
+		logger.Info("Found genesis file", "path", genFile)
+	} else {
+		genDoc := comettypes.GenesisDoc{
+			ChainID:         fmt.Sprintf("test-rollup-%08x", rand.Uint32()), //nolint:gosec
+			GenesisTime:     time.Now().Round(0).UTC(),
+			ConsensusParams: comettypes.DefaultConsensusParams(),
+		}
+		pubKey, err := pv.GetPubKey()
+		if err != nil {
+			return fmt.Errorf("can't get pubkey: %w", err)
+		}
+		genDoc.Validators = []comettypes.GenesisValidator{{
+			Address: pubKey.Address(),
+			PubKey:  pubKey,
+			Power:   1000,
+			Name:    "Rollkit Sequencer",
+		}}
+
+		if err := genDoc.SaveAs(genFile); err != nil {
+			return err
+		}
+		logger.Info("Generated genesis file", "path", genFile)
+	}
+
+	return nil
+}
