+++ cmd/rollkit/commands/run_node.go
@@ -3,7 +3,6 @@ package commands
 import (
 	"context"
 	"fmt"
-	"math/rand"
 	"os"
 	"time"
 
@@ -14,7 +13,6 @@ import (
 	cometnode "github.com/cometbft/cometbft/node"
 	cometp2p "github.com/cometbft/cometbft/p2p"
 	cometprivval "github.com/cometbft/cometbft/privval"
-	comettypes "github.com/cometbft/cometbft/types"
 	"github.com/mitchellh/mapstructure"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
@@ -38,21 +36,6 @@ func NewRunNodeCmd() *cobra.Command {
 		Use:     "start",
 		Aliases: []string{"node", "run"},
 		Short:   "Run the rollkit node",
-		// PersistentPreRunE is used to parse the config and initial the config files
-		PersistentPreRunE: func(cmd *cobra.Command, args []string) error {
-			err := parseConfig(cmd)
-			if err != nil {
-				return err
-			}
-
-			// use aggregator by default if the flag is not specified explicitly
-			if !cmd.Flags().Lookup("rollkit.aggregator").Changed {
-				nodeConfig.Aggregator = true
-			}
-
-			// Initialize the config files
-			return initFiles(log.NewNopLogger()) //TODO: place the correct logger
-		},
 		RunE: func(cmd *cobra.Command, args []string) error {
 			genDocProvider := cometnode.DefaultGenesisDocProviderFunc(config)
 			genDoc, err := genDocProvider()
@@ -196,65 +179,6 @@ func addNodeFlags(cmd *cobra.Command) {
 
 	// Add Rollkit flags
 	rollconf.AddFlags(cmd)
-}
-
-// TODO (Ferret-san): modify so that it initiates files with rollkit configurations by default
-// note that such a change would also require changing the cosmos-sdk
-func initFiles(logger log.Logger) error {
-	// Generate the private validator config files
-	cometprivvalKeyFile := config.PrivValidatorKeyFile()
-	cometprivvalStateFile := config.PrivValidatorStateFile()
-	var pv *cometprivval.FilePV
-	if fileExists(cometprivvalKeyFile) {
-		pv = cometprivval.LoadFilePV(cometprivvalKeyFile, cometprivvalStateFile)
-		logger.Info("Found private validator", "keyFile", cometprivvalKeyFile,
-			"stateFile", cometprivvalStateFile)
-	} else {
-		pv = cometprivval.GenFilePV(cometprivvalKeyFile, cometprivvalStateFile)
-		pv.Save()
-		logger.Info("Generated private validator", "keyFile", cometprivvalKeyFile,
-			"stateFile", cometprivvalStateFile)
-	}
-
-	// Generate the node key config files
-	nodeKeyFile := config.NodeKeyFile()
-	if fileExists(nodeKeyFile) {
-		logger.Info("Found node key", "path", nodeKeyFile)
-	} else {
-		if _, err := cometp2p.LoadOrGenNodeKey(nodeKeyFile); err != nil {
-			return err
-		}
-		logger.Info("Generated node key", "path", nodeKeyFile)
-	}
-
-	// Generate the genesis file
-	genFile := config.GenesisFile()
-	if fileExists(genFile) {
-		logger.Info("Found genesis file", "path", genFile)
-	} else {
-		genDoc := comettypes.GenesisDoc{
-			ChainID:         fmt.Sprintf("test-rollup-%08x", rand.Uint32()), //nolint:gosec
-			GenesisTime:     time.Now().Round(0).UTC(),
-			ConsensusParams: comettypes.DefaultConsensusParams(),
-		}
-		pubKey, err := pv.GetPubKey()
-		if err != nil {
-			return fmt.Errorf("can't get pubkey: %w", err)
-		}
-		genDoc.Validators = []comettypes.GenesisValidator{{
-			Address: pubKey.Address(),
-			PubKey:  pubKey,
-			Power:   1000,
-			Name:    "Rollkit Sequencer",
-		}}
-
-		if err := genDoc.SaveAs(genFile); err != nil {
-			return err
-		}
-		logger.Info("Generated genesis file", "path", genFile)
-	}
-
-	return nil
 }
 
 func parseConfig(cmd *cobra.Command) error {
